リファクタリングの実施
