+++ src/index.template.html
@@ -8,7 +8,18 @@
     <meta name="format-detection" content="telephone=no">
     <meta name="msapplication-tap-highlight" content="no">
     <meta name="viewport" content="user-scalable=no, initial-scale=1, maximum-scale=1, minimum-scale=1, width=device-width<% if (ctx.mode.cordova || ctx.mode.capacitor) { %>, viewport-fit=cover<% } %>">
-
+    <!-- OGP -->
+    <meta property="og:title" content="<%= productName %>" />
+    <meta
+      property="og:description"
+      content="<%= productDescription %>">
+    />
+    <meta property="og:type" content="website" />
+    <meta property="og:url" content="https://yama-code.web.app" />
+    <meta
+      property="og:image"
+      content="https://yama-code.web.app/yama-icon.jpg"
+    />
     <link rel="icon" type="image/jpg" sizes="128x128" src="yama-icon.jpg">
     <link rel="icon" type="image/jpg" sizes="96x96" src="yama-icon.jpg">
     <link rel="icon" type="image/jpg" sizes="32x32" src="yama-icon.jpg">
