+++ src/pages/Index.vue
@@ -1,148 +1,8 @@
 <template>
   <q-page-container>
-    <q-page class="padding">
-      <div class="inner">
-        <br />
-        <div class="text-h3">
-          Who is yama?
-        </div>
-        <br /><br />
-        <div class="icon-image">
-          <q-img
-            src="https://pbs.twimg.com/profile_images/986113842739478528/2PlaIAsz_400x400.jpg"
-            style="height: 280px; max-width: 300px"
-          ></q-img>
-        </div>
-        <br />
-        <div class="text-h4">
-          yama
-        </div>
-        <br />
-      </div>
-      <div
-        class="layout-padding row justify-center background-whitesmoke items-center"
-      >
-        <div class="col-xs-10 col-sm-7 col-lg-5">
-          <div
-            class="layout-padding row justify-center background-whitesmoke items-center"
-          >
-            <q-card flat bordered class="my-card">
-              <q-card-section>
-                <div class="text-h6">Engineer & Creator</div>
-              </q-card-section>
-
-              <q-card-section class="q-pt-none">
-                <p v-html="parser.translateHTMLString('東京都立大学大学院 システムデザイン研究科 情報科学域 博士前期過程')"></p>
-              </q-card-section>
-              <q-separator inset />
-              <q-card-section>
-                <q-avatar size="30px">
-                  <img
-                    src="https://cdn.icon-icons.com/icons2/936/PNG/128/github-logo_icon-icons.com_73546.png"
-                  /> </q-avatar
-                >&nbsp;&nbsp;
-                <a href="https://github.com/kurikinton105">kurikinton105</a
-                ><br /><br />
-                <q-avatar size="30px">
-                  <img src="~assets/Twitter_Logo_WhiteOnBlue.png" /> </q-avatar
-                >&nbsp;&nbsp;
-                <a href="https://twitter.com/y_a_m_a_y_a">@y_a_m_a_y_a</a
-                ><br /><br />
-                <q-avatar size="30px">
-                  <img src="~assets/qiita-qiita-jobs-favicon.png" /> </q-avatar
-                >&nbsp;&nbsp; <a href="https://qiita.com/y_a_m_a">y_a_m_a</a
-                ><br /><br />
-                <q-avatar size="30px">
-                  <img src="~assets/hatenablog-logo.png" /> </q-avatar
-                >&nbsp;&nbsp;
-                <a href="https://y-a-m-a-y-a.hatenablog.com/"
-                  >yamaのメタノート</a
-                ><br /><br />
-                <q-avatar square size="30px">
-                  <img src="https://content.linkedin.com/content/dam/me/brand/en-us/brand-home/logos/In-Blue-Logo.png.original.png" /> </q-avatar
-                >&nbsp;&nbsp;
-                <a href="https://www.linkedin.com/in/yamada-kenta5/"
-                  >yamada-kenta5</a
-                ><br /><br />
-              </q-card-section>
-            </q-card>
-          </div>
-        </div>
-      </div>
-      <br /><br />
-    </q-page>
-    <hr />
-    <q-page>
-      <br />
-      <div class="q-px-lg q-pb-md">
-        <q-timeline :layout="layout" color="secondary">
-          <q-timeline-entry heading>
-            Activity Record
-          </q-timeline-entry>
-          <div>
-            <q-timeline-entry
-              v-for="Activity in Activities.data"
-              :key="Activity"
-              v-bind:title="Activity.name"
-              v-bind:subtitle="Activity.date_info"
-              v-bind:side="Activity.side"
-              v-bind:icon="Activity.icon"
-            >
-              <div>
-                <p v-html="parser.translateHTMLString(Activity.infomation)"></p>
-              </div>
-            </q-timeline-entry>
-          </div>
-        </q-timeline>
-      </div>
-    </q-page>
-    <hr />
-    <q-page>
-      <div class="inner">
-        <br />
-        <div class="text-h3">
-          Articles
-        </div>
-        <br />
-        <div class="text-weight-regular">
-          Qiita、はてなブログなどから記事の一覧をまとめています。
-        </div>
-      </div>
-      <br />
-      <div class="justify-evenly">
-        <div class="flex flex-center ">
-          <div class="row justify-center q-gutter-sm">
-            <div v-for="Article in Articles.data" :key="Article">
-              <div class="col-auto">
-                <q-card class="my-card-artcle" flat bordered>
-                  <br />
-                  <q-item>
-                    <q-item-section avatar>
-                      <q-avatar size="30px">
-                        <img :src="Article.site" />
-                      </q-avatar>
-                    </q-item-section>
-
-                    <q-item-section>
-                      <q-item-label
-                        ><a v-bind:href="Article.link" target="_blank">{{
-                          Article.name
-                        }}</a></q-item-label
-                      >
-                      <q-item-label caption>
-                        {{ Article.date }}
-                      </q-item-label>
-                    </q-item-section>
-                  </q-item>
-                  <br />
-                </q-card>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    </q-page>
-
+    <Info /><hr/>
+    <Activity /><hr />
+    <Article />
     <q-page-scroller
       position="bottom-right"
       :scroll-offset="150"
@@ -153,36 +13,14 @@
   </q-page-container>
 </template>
 
-<style lang="sass" scoped>
-.my-card-artcle
-  width: 100%
-  min-width :400px
-  max-width: 400px
-  height:100px
-.inner
-  text-align: center
-.bluepage
-  background-color: #0000
-</style>
 <script src="https://www.gstatic.com/firebasejs/4.1.3/firebase.js"></script>
 <script>
-import axios from "axios";
-import { VueTyper } from "vue-typer";
-import { loadDefaultJapaneseParser } from "budoux";
-import { defineComponent } from "vue";
+import Info from "src/components/Info.vue";
+import Activity from "src/components/Activity.vue";
+import Article from "src/components/Articles.vue";
 
 export default {
   name: "PageIndex",
-  data() {
-    const parser = loadDefaultJapaneseParser(); // budoux
-    return {
-      parser,
-      allDataID: "non",
-      Activities: "non",
-      Articles: "non",
-      Skills: "non"
-    };
-  },
   computed: {
     layout() {
       return this.$q.screen.lt.sm
@@ -193,28 +31,7 @@ export default {
     }
   },
   components: {
-    VueTyper
-  },
-  mounted: function() {
-    axios
-      .get(
-        "https://us-central1-cos5year.cloudfunctions.net/SheetsAPI?sheet_id=1FeRqAE3YLWKfvHomN23ANyZONzT0hKBtAFp08H-c9gI"
-      )
-      .then(response => (this.allDataID = response));
-
-    //console.log('check sample_url')
-    //this.allDataID="bbb"
-    //console.log(this.allDataID)
-    //console.log('check')
-    axios
-      .get("https://api.sssapi.app/Dg3gg3OE7uDaEbBzq2DPb")
-      .then(response => (this.Activities = response));
-
-    axios
-      .get("https://api.sssapi.app/V-sFi3aRWS1Ykb1j3m2qO")
-      .then(response => (this.Articles = response));
-    console.log(this.allDataID);
-    console.log(this.data.Activity);
+    Info,Activity,Article
   }
 };
 </script>
